Example NEMA: generate graph even if exception in the run of the model Example Scenarios_monoculms : rename scenarii by scenarios + minor changes
