--- example/Scenarii_monoculms/inputs/Ljutovac2002.csv
@@ -1,12 +1,0 @@
-metamer,leaf_Lmax,lamina_Lmax,sheath_Lmax,internode_Lmax,lamina_Wmax,lamina_area,facteurForme_lamina,RER,RER_confint
-1,11.066,8.213,2.853,,0.3,1.883,0.75,,
-2,12.104,9.231,2.873,,0.33,2.152,0.71,,
-3,12.277,9.412,2.865,,0.4,2.78,0.73,,
-4,13.314,9.987,3.327,,0.45,3.318,0.74,,
-5,15.476,11.398,4.077,,0.56,4.664,0.73,2.79E-06,3.35E-07
-6,19.971,13.646,6.326,0.101,0.75,7.713,0.75,1.76E-06,2.07E-07
-7,25.504,16.434,9.071,1.913,1.01,12.287,0.74,1.62E-06,8.33E-08
-8,31.816,19.665,12.151,6.091,1.21,17.758,0.75,1.44E-06,7.73E-08
-9,39.164,25.108,14.056,9.715,1.28,24.215,0.75,1.44E-06,5.56E-08
-10,45.994,28.683,17.311,14.497,1.43,30.762,0.75,1.44E-06,5.40E-08
-11,42.709,23.947,18.762,17.013,1.8,32.556,0.75,1.42E-06,4.61E-08
