--- example/Scenarii_monoculms/inputs/axeTable.csv
@@ -1,2 +1,0 @@
-HS_final,axe,azT,azTb,disp,dredT,emf1,end,hasEar,incT,nf,nf_end,plant
-12,MS,158.4467164,0,NA,0,21.54309661,NA,1,1.599403718,12,12,1
