--- example/Scenarii_monoculms/inputs/soils_initial_state.csv
@@ -1,2 +1,0 @@
-plant,axis,volume,nitrates,Tsoil
-1,MS,1,0,12
