+++ example/Scenarios_monoculms/inputs/hiddenzones_initial_state.csv
@@ -1,0 +1,9 @@
+plant,axis,metamer,hiddenzone_age,leaf_pseudo_age,internode_pseudo_age,leaf_is_growing,internode_is_growing,leaf_pseudostem_length,internode_distance_to_emerge,leaf_L,internode_L,delta_leaf_pseudo_age,delta_internode_pseudo_age,delta_internode_distance_to_emerge,delta_internode_L,delta_leaf_L,delta_leaf_pseudostem_length,leaf_Lmax,lamina_Lmax,sheath_Lmax,leaf_Wmax_int,leaf_Wmax,SSLW,LSSW,leaf_is_emerged,internode_Lmax,LSIW,internode_is_visible,sucrose,amino_acids,fructan,proteins,leaf_enclosed_mstruct,internode_enclosed_mstruct,mstruct,leaf_enclosed_Nstruct,internode_enclosed_Nstruct,Nstruct,mean_conc_sucrose,is_over,leaf_is_remobilizing,internode_is_remobilizing,leaf_Lmax_em,internode_Lmax_lig,ratio_DZ
+1,MS,3,3835440,1570000,864000,True,True,0.02895,0.0289,0.0925,0.00005,3600,3600,0,0,0,0,0.1228,0.093426659,0.029373341,4.63E-05,0.004,21,9.00E-02,True,NA,1,False,8.908843665,0.996526223,8.908843665,9.965262232,0.002018013,1.32E-06,0.002019338,1.01E-05,6.62E-09,1.01E-05,1.80E+03,False,False,False,1.23E-01,NA,1
+1,MS,4,3287520,757000,0,True,True,0.02894,0.02893,0.0265,0.00001,3600,0,0,0,0,0,0.1331,0.099106478,0.033993522,2.68E-05,NA,NA,NA,False,NA,NA,False,3.588122085,0.675836276,1.794061043,6.758362756,0.000813139,1.69E-07,0.000813308,4.07E-06,8.44E-10,4.07E-06,1.80E+03,False,False,False,1.33E-01,NA,1
+1,MS,5,2739600,0,0,True,False,0.02894,0.02894,0.001967739,0,0,0,0,0,0,0,NA,NA,NA,1.55E-05,NA,NA,NA,False,NA,NA,False,0.128730408,0.053069049,0,0.530690492,2.92E-05,0,2.92E-05,1.46E-07,0.00E+00,1.46E-07,1.80E+03,False,False,False,NA,NA,1
+1,MS,6,2191680,0,0,True,False,0.02894,0.02894,0.000754112,0,0,0,0,0,0,0,NA,NA,NA,8.95E-06,NA,NA,NA,False,NA,NA,False,0.037798204,0.015582292,0,0.155822917,8.57E-06,0,8.57E-06,4.28E-08,0.00E+00,4.28E-08,NA,False,False,False,NA,NA,1
+1,MS,7,1643760,0,0,True,False,0.02894,0.02894,0.000313764,0,0,0,0,0,0,0,NA,NA,NA,5.17E-06,NA,NA,NA,False,NA,NA,False,0.012381692,0.005104347,0,0.051043466,2.81E-06,0,2.81E-06,1.40E-08,0.00E+00,1.40E-08,NA,False,False,False,NA,NA,1
+1,MS,8,1095840,0,0,True,False,0.02894,0.02894,0.000145874,0,0,0,0,0,0,0,NA,NA,NA,2.99E-06,NA,NA,NA,False,NA,NA,False,0.0047184,0.001945158,0,0.019451581,1.07E-06,0,1.07E-06,5.35E-09,0.00E+00,5.35E-09,NA,False,False,False,NA,NA,1
+1,MS,9,547920,0,0,True,False,0.02894,0.02894,7.80E-05,0,0,0,0,0,0,0,NA,NA,NA,1.73E-06,NA,NA,NA,False,NA,NA,False,0.002181585,0.000899357,0,0.008993572,4.94E-07,0,4.94E-07,2.47E-09,0.00E+00,2.47E-09,NA,False,False,False,NA,NA,1
+1,MS,10,0,0,0,True,False,0.02894,0.02894,4.94E-05,0,0,0,0,0,0,0,NA,NA,NA,0.000001,NA,NA,NA,False,NA,NA,False,0.001266463,0.000522099,0,0.005220988,2.87E-07,0,2.87E-07,1.44E-09,0,1.44E-09,NA,False,False,False,NA,NA,1
