+++ example/Scenarios_monoculms/inputs/organs_initial_state.csv
@@ -1,0 +1,3 @@
+plant,axis,organ,sucrose,amino_acids,nitrates,structure,starch,proteins,mstruct,senesced_mstruct,Nstruct,cytokinins
+1,MS,phloem,0,0,NA,NA,NA,NA,NA,NA,NA,NA
+1,MS,roots,202.5847905,2.19053373,21.9053373,NA,NA,NA,3.44E-02,0.00E+00,0.000172197,8.60E+00
