+++ example/Scenarios_monoculms/inputs/soils_initial_state.csv
