+++ example/Scenarios_monoculms/inputs_Ljutovac_N2/elements_initial_state.csv
