+++ example/Scenarios_monoculms/inputs_Ljutovac_N2/hiddenzones_initial_state.csv
