+++ example/Scenarios_monoculms/inputs_Ljutovac_N2/soils_initial_state.csv
@@ -1,0 +1,2 @@
+plant,axis,volume,nitrates,Tsoil
+1,MS,1,0,12
